// src/api/types.gen.ts
export type UserResponse = {
  /**
   * Id
   */
  id: string
  /**
   * Email
   */
  email: string
  /**
   * Is Active
   */
  is_active: boolean
  /**
   * Is Superuser
   */
  is_superuser: boolean
  /**
   * Is Verified
   */
  is_verified: boolean
  /**
   * Created At
   */
  created_at: string
  /**
   * Updated At
   */
  updated_at?: string | null
}